requirements.txt
requests
reportlab
pycups
google-play-scraper
python-dotenv
actualpy
PyYAML
qrcode[pil]
